Moving map logic to its dedicaced file
